feat: add Feature Importance Ranking lollipop chart to Explore page
- New FeatureImportanceChart component with interactive lollipop chart
- Uses official WEKA Information Gain values from research paper
- Category filter pills (Institutional, Demographic, Digital, Economic)
- Hover interactions with color-coded stems and dots
- Fixed-height container to prevent layout shift on filter change
- Integrated below the map on ExplorePage

+++ backend/app/api/counties.py
@@ -46,7 +46,7 @@ def counties_statistics(db: Session = Depends(get_db)):
     without_programs = crud_counties.get_counties_count(db, has_programs=0)
     counties = crud_counties.get_counties(db, limit=10000)
     avg_impact = (
-        sum(c.total_program_impact_score for c in counties) / len(counties)
+        float(sum(c.total_program_impact_score for c in counties) / len(counties))
         if counties else 0
     )
     highest = max(counties, key=lambda c: c.total_program_impact_score, default=None)
